航票系统实体类和form下面的扩展类

+++ we-web-parent/web-linyun-airline/src/main/java/com/linyun/airline/forms/DictRelationAddForm.java
@@ -1,20 +1,21 @@
 package com.linyun.airline.forms;
 
+import java.io.Serializable;
+
 import lombok.Data;
 import lombok.EqualsAndHashCode;
-import com.uxuexi.core.web.form.AddForm;
 
-import java.io.Serializable;
+import com.uxuexi.core.web.form.AddForm;
 
 @Data
 @EqualsAndHashCode(callSuper = true)
-public class DictRelationAddForm extends AddForm implements Serializable{
+public class DictRelationAddForm extends AddForm implements Serializable {
 	private static final long serialVersionUID = 1L;
-		
+
 	/**字典id*/
-	private Integer sourceId;
-		
+	private long sourceId;
+
 	/**关联字典id*/
-	private Integer targetId;
-		
+	private long targetId;
+
 }
