+++ we-web-parent/web-linyun-airline/src/main/java/com/linyun/airline/forms/DictRelationUpdateForm.java
@@ -1,20 +1,21 @@
 package com.linyun.airline.forms;
 
+import java.io.Serializable;
+
 import lombok.Data;
 import lombok.EqualsAndHashCode;
-import com.uxuexi.core.web.form.ModForm;
 
-import java.io.Serializable;
+import com.uxuexi.core.web.form.ModForm;
 
 @Data
 @EqualsAndHashCode(callSuper = true)
-public class DictRelationUpdateForm extends ModForm implements Serializable{
+public class DictRelationUpdateForm extends ModForm implements Serializable {
 	private static final long serialVersionUID = 1L;
-		
+
 	/**字典id*/
-	private Integer sourceId;
-		
+	private long sourceId;
+
 	/**关联字典id*/
-	private Integer targetId;
-		
+	private long targetId;
+
 }
